Move exception classes to own namespace
