--- src/main/java/com/keboola/tableexporter/ApplicationException.java
@@ -1,7 +1,0 @@
-package com.keboola.tableexporter;
-
-public class ApplicationException extends Exception {
-    public ApplicationException(String message, Throwable previous) {
-        super(message, previous);
-    }
-}
