--- src/main/java/com/keboola/tableexporter/UserException.java
@@ -1,7 +1,0 @@
-package com.keboola.tableexporter;
-
-public class UserException extends Exception {
-    public UserException(String message, Throwable previous) {
-        super(message, previous);
-    }   
-}
