+++ src/main/java/com/keboola/tableexporter/exception/ApplicationException.java
@@ -1,0 +1,7 @@
+package com.keboola.tableexporter.exception;
+
+public class ApplicationException extends Exception {
+    public ApplicationException(String message, Throwable previous) {
+        super(message, previous);
+    }
+}
