+++ src/main/java/com/keboola/tableexporter/exception/CsvException.java
@@ -1,4 +1,4 @@
-package com.keboola.tableexporter;
+package com.keboola.tableexporter.exception;
 
 public class CsvException extends Exception {
     public CsvException(String message, Throwable previous) {
