+++ src/main/java/com/keboola/tableexporter/exception/UserException.java
@@ -1,0 +1,7 @@
+package com.keboola.tableexporter.exception;
+
+public class UserException extends Exception {
+    public UserException(String message, Throwable previous) {
+        super(message, previous);
+    }   
+}
